pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ark-agent-core"
version = "0.1.0"
description = "智能助理框架：Skill 系統 + Workflow 引擎 + LLM 整合"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Paddy Yang"},
]
keywords = ["ai-agent", "skill", "workflow", "llm", "wiki"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
mongodb = ["pymongo>=4.0"]
mssql = ["pymssql>=2.2"]
bigquery = ["google-cloud-bigquery>=3.0"]
telegram = ["python-telegram-bot[ext]>=21.0"]
all = [
    "pymongo>=4.0",
    "pymssql>=2.2",
    "google-cloud-bigquery>=3.0",
    "python-telegram-bot[ext]>=21.0",
]

[project.scripts]
ark = "ark_agent_core.cli.main:main"

[tool.setuptools.packages.find]
include = ["ark_agent_core*"]
